requirements_local.txt
asgiref==3.5.2
autopep8==1.7.0
cffi==1.15.1
cryptography==37.0.4
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
pycodestyle==2.9.1
pycparser==2.21
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
tzdata==2022.1
